machine translation projects/Seq2SeqAttention/requirements.txt
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
matplotlib==3.1.1
numpy==1.16.3
scikit-learn==0.21.3
tensorboard==2.0.1
tensorflow-estimator==2.0.1
tensorflow-gpu==2.0.0

